[FE] rename, fix: game 위치 수정 및 오브젝트 수정
