+++ frontend/src/utils/game/game.js
@@ -1,0 +1,89 @@
+import * as THREE from 'three';
+import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
+import createMap from './createMap';
+
+function game() {
+  const canvas = document.getElementById('gameCanvas');
+  const renderer = new THREE.WebGLRenderer({ antialias: true, canvas });
+
+  const fov = 45;
+  const aspect = 2; // the canvas default
+  const near = 0.1;
+  const far = 1000;
+  const camera = new THREE.PerspectiveCamera(fov, aspect, near, far);
+  camera.position.set(0, 0, 100);
+
+  const controls = new OrbitControls(camera, canvas);
+  controls.target.set(0, 0, 0);
+  controls.update();
+
+  const scene = new THREE.Scene();
+
+  // 배경 이미지
+  // const loader = new THREE.TextureLoader();
+  // loader.load('/img/map_pixel_rain.jpg', function (texture) {
+  //   scene.background = texture;
+  // });
+  scene.background = new THREE.Color(0x000000);
+
+  {
+    const skyColor = 0xb1e1ff; // light blue
+    const groundColor = 0xb97a20; // brownish orange
+    const intensity = 2;
+    const light = new THREE.HemisphereLight(skyColor, groundColor, intensity);
+    scene.add(light);
+  }
+
+  {
+    const color = 0xffffff;
+    const intensity = 2.5;
+    const light = new THREE.DirectionalLight(color, intensity);
+    light.position.set(5, 10, 2);
+    scene.add(light);
+    scene.add(light.target);
+  }
+
+  createMap(scene);
+
+  const redPaddleGeometry = new THREE.BoxGeometry(0.5, 3, 3);
+  const redPaddleMaterial = new THREE.MeshPhongMaterial({ color: 0xff0000 });
+  const redPaddle = new THREE.Mesh(redPaddleGeometry, redPaddleMaterial);
+  redPaddle.position.set(25, 0, 0);
+  scene.add(redPaddle);
+
+  const bluePaddleGeometry = new THREE.BoxGeometry(0.5, 3, 3);
+  const bluePaddleMaterial = new THREE.MeshPhongMaterial({ color: 0x0000ff });
+  const bluePaddle = new THREE.Mesh(bluePaddleGeometry, bluePaddleMaterial);
+  bluePaddle.position.set(-25, 0, 0);
+  scene.add(bluePaddle);
+
+
+  function resizeRendererToDisplaySize(renderer) {
+    const canvas = renderer.domElement;
+    const width = canvas.clientWidth;
+    const height = canvas.clientHeight;
+    const needResize = canvas.width !== width || canvas.height !== height;
+    renderer.setPixelRatio(window.devicePixelRatio);
+    if (needResize) {
+      renderer.setSize(width, height, false);
+    }
+    return needResize;
+  }
+
+  function render(time) {
+    time *= 0.001;
+
+    if (resizeRendererToDisplaySize(renderer)) {
+      const canvas = renderer.domElement;
+      camera.aspect = canvas.clientWidth / canvas.clientHeight;
+      camera.updateProjectionMatrix();
+    }
+    renderer.render(scene, camera);
+
+    requestAnimationFrame(render);
+  }
+
+  requestAnimationFrame(render);
+}
+
+export default game;
